prepare for testing
